:technologist: Scheduler has only attribute 'Status' and 'RuleApplied' Delete all the specific rule guards for the components because they are not really necessary. We only have to keep track whether a rule of the currently necessary component has already been applied. Thus, we only need one rule guard attribute.
If a rule in the currently necessary component is applied, then the 
'RuleApplied' attribute is set to 'true'. The manager then sets the 
'Status' to the next component and the 'RuleApplied' attribute to 
'false.

A special case are the quality attributes because we have to compute 
three values here. In their rules, they don't change the value of the 
'RuleApplied' attribute. Instead, there is an equation in the manager 
that sets the 'RuleApplied' attribute to 'true' if all QAs have been 
computed.